fix: move Overlay.Footer out of the Overlay.Body
