+++ src/dataLoaders/components/DataLoadersOverlay.scss
@@ -36,10 +36,6 @@
   }
 }
 
-.plugin-create-configuration {
-  height: 100%;
-}
-
 .data-loading--column-panel {
   background-color: $g3-castle;
   border-radius: 0 $radius $radius 0;
